+++ src/Stripe.net/Entities/Customers/Customer.cs
@@ -17,7 +17,7 @@ namespace Stripe
         /// Current balance, if any, being stored on the customer’s account. If negative, the customer has credit to apply to the next invoice. If positive, the customer has an amount owed that will be added to the next invoice. The balance does not refer to any unpaid invoices; it solely takes into account amounts that have yet to be successfully applied to any invoice. This balance is only taken into account for recurring billing purposes (i.e., subscriptions, invoices, invoice items)
         /// </summary>
         [JsonProperty("account_balance")]
-        public int AccountBalance { get; set; }
+        public long AccountBalance { get; set; }
 
         [JsonProperty("created")]
         public DateTime Created { get; set; }
